requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
requests==2.32.3
chromadb==0.5.4
reportlab==4.2.2
